pyproject.toml
[tool.poetry]
name = "autolabmate"
version = "0.1.0"
description = "Web app for Bayesian optimization aimed at researchers with expertise in disciplines other than statistics and computer science."
authors = ["Hud Wahab <hudwahab@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9.7,<3.12"
rpy2 = "^3.5.14"
pandas = "^2.1.2"
flask = "^3.0.0"
streamlit = "1.35.0"
watchdog = "^3.0.0"
plotly = "^5.18.0"
fastapi = "^0.104.0"
supabase = "^1.2.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
gotrue = "^1.2.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.22"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
streamlit-extras = "^0.3.4"
st-pages = "^0.4.5"
scikit-learn = "^1.3.2"
pdpbox = "^0.3.0"
pydoe = "^0.3.8"
sobol-seq = "^0.2.0"
simplejson = "^3.19.2"
pytest-mock = "^3.14.0"
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
